refactor(anytype): extract collection/property logic into collection.ts
Move the collection and property functionality out of the growing importer.ts
into a dedicated collection.ts (mirroring the Notion importer): the relation /
option / file-object indexers, property and collection parsers, the column
helpers, and the note mutators (createCollectionNote / applyProperties /
applyFiles), plus their types. importer.ts keeps the zip reading, page
predicates, block rendering and the note-tree builder, importing from
collection.ts; the only runtime dependency is importer -> collection (collection
imports the base model types type-only).

Relocate the collection-properties unit tests to collection.spec.ts with their
assertions unchanged; the end-to-end collection cases stay in the importer
integration spec. importer.ts drops from ~1130 to ~780 lines.
